fix: AutomaticBackupDisabledOnAppUpdate
